refactor: InputSection 네이밍, 폴더 구조 수정
