+++ src/components/Box/InputSection/CardNumberInput.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import InputSection from '../Box/InputSection';
-import InputBox from '../Common/InputBox';
-import { InputStateProps } from '../../types';
+import InputSectionTemplate from '../InputSectionTemplate';
+import InputBox from '../../Common/InputBox';
+import { InputStateProps } from '../../../types';
 
 const CardNumberInput = (props: InputStateProps) => {
   const inputs = [
@@ -11,9 +11,9 @@ const CardNumberInput = (props: InputStateProps) => {
     { type: 'number', maxLength: 4, placeholder: '∙∙∙∙', required: true, textSecurity: true },
   ];
   return (
-    <InputSection label="카드 번호">
+    <InputSectionTemplate label="카드 번호">
       <InputBox inputs={inputs} align="center" separator="-" isFullWidth {...props} />
-    </InputSection>
+    </InputSectionTemplate>
   );
 };
