+++ src/components/Box/InputSectionTemplate.tsx
